+++ R/EngineGraphQLGitLab.R
@@ -95,6 +95,7 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
                             settings) {
        NULL
      }
+
    ),
 
    private = list(
@@ -167,6 +168,8 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
        if (length(repos_list) > 0) {
          repos_table <- purrr::map_dfr(repos_list, function(repo) {
            repo <- repo$node
+           repo$default_branch <- repo$repository$rootRef
+           repo$repository <- NULL
            repo$languages <- if (length(repo$languages) > 0) {
              purrr::map_chr(repo$languages, ~ .$name) %>%
                paste0(collapse = ", ")
@@ -185,6 +188,10 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
            dplyr::relocate(
              repo_url,
              .after = organization
+           ) %>%
+           dplyr::relocate(
+             default_branch,
+             .after = name
            )
        } else {
          repos_table <- NULL
@@ -214,6 +221,74 @@ EngineGraphQLGitLab <- R6::R6Class("EngineGraphQLGitLab",
          user_table <- NULL
        }
        return(user_table)
+     },
+
+     # @description Pull all given files from all repositories of a group.
+     # @param org An organization.
+     # @param file_path Path to a file.
+     # @return A response in a list form.
+     pull_file_from_org = function(org, file_path) {
+       full_files_list <- list()
+       next_page <- TRUE
+       end_cursor <- ""
+       while (next_page) {
+         files_query <- self$gql_query$files_by_org(
+           end_cursor = end_cursor
+         )
+         files_response <- self$gql_response(
+           gql_query = files_query,
+           vars = list(
+             "org" = org,
+             "file_paths" = file_path
+           )
+         )
+         if (length(files_response$data$group) == 0) {
+           cli::cli_abort("Empty")
+         }
+         projects <- files_response$data$group$projects
+         files_list <- purrr::map(projects$edges, function(edge) {
+           edge$node
+         }) %>%
+           purrr::discard(~ length(.$repository$blobs$nodes) == 0)
+         if (is.null(files_list)) files_list <- list()
+         if (length(files_list) > 0) {
+           next_page <- files_response$pageInfo$hasNextPage
+         } else {
+           next_page <- FALSE
+         }
+         if (is.null(next_page)) next_page <- FALSE
+         if (next_page) {
+           end_cursor <- files_response$pageInfo$endCursor
+         } else {
+           end_cursor <- ""
+         }
+         full_files_list <- append(full_files_list, files_list)
+       }
+       return(full_files_list)
+     },
+
+     # @description Prepare files table.
+     # @param files_response A list.
+     # @param org An organization.
+     # @return A table with information on files.
+     prepare_files_table = function(files_response, org, file_path) {
+       if (!is.null(files_response)) {
+         files_table <- purrr::map(files_response, function(project) {
+           data.frame(
+             "repository_name" = project$name,
+             "repository_id" = project$id,
+             "organization" = org,
+             "file_path" = project$repository$blobs$nodes[[1]]$name,
+             "file_content" = project$repository$blobs$nodes[[1]]$rawBlob,
+             "file_size" = as.integer(project$repository$blobs$nodes[[1]]$size),
+             "api_url" = self$gql_api_url
+           )
+         }) %>%
+           purrr::list_rbind()
+       } else {
+         files_table <- NULL
+       }
+       return(files_table)
      }
    )
 )