+++ tests/testthat/test-03-EngineGraphQLGitHub.R
@@ -206,6 +206,26 @@ test_that("GitHub prepares user table", {
   test_mocker$cache(gh_user_table)
 })
 
+test_that("GitHub GraphQL Engine pulls files from organization", {
+  expect_snapshot(
+    github_files_response <- test_gql_gh$pull_file_from_org(
+      "r-world-devs",
+      "meta_data.yaml"
+    )
+  )
+  expect_github_files_response(github_files_response)
+  test_mocker$cache(github_files_response)
+})
+
+test_that("GitHub GraphQL Engine prepares table from files response", {
+  files_table <- test_gql_gh$prepare_files_table(
+    files_response = test_mocker$use("github_files_response"),
+    org = "r-world-devs",
+    file_path = "meta_data.yaml"
+  )
+  expect_files_table(files_table)
+})
+
 # public methods
 
 test_gql_gh <- EngineGraphQLGitHub$new(
@@ -286,4 +306,15 @@ test_that("`pull_commits()` retrieves commits in the table format", {
   expect_commits_table(
     commits_table
   )
+})
+
+test_that("`pull_files()` pulls files in the table format", {
+  expect_snapshot(
+    gh_files_table <- test_gql_gh$pull_files(
+      org = "r-world-devs",
+      file_path = "LICENSE"
+    )
+  )
+  expect_files_table(gh_files_table)
+  test_mocker$cache(gh_files_table)
 })